+++ lib/analog_radio.dart
@@ -1,4 +1,3 @@
 library analog_radio;
 
 export 'src/analog_radio.dart';
-export 'src/filter_type.dart';