+++ lib/src/analog_radio.dart
@@ -1,11 +1,14 @@
+import 'dart:web_audio';
+
 import 'audio_processor.dart';
 import 'audio_processor_state.dart';
-import 'filter_type.dart';
 
 /// Emulates analog radio
 abstract class AnalogRadio {
-  factory AnalogRadio() {
-    return AudioProcessor();
+  factory AnalogRadio({
+    AudioNode Function(AudioContext context)? setCustomNode,
+  }) {
+    return AudioProcessor(setCustomNode: setCustomNode);
   }
 
   /// Radio state
@@ -20,6 +23,9 @@ abstract class AnalogRadio {
   /// Adjusts the radio to the [url] stream with the signal strength [signalStrength]. [signalStrength] must be from 0 to 1
   void tune(String url, double signalStrength);
 
+  /// Turns on the frequency change sound
+  void tuning();
+
   /// The current volume. From 0 to 1
   double volume = 1;
 
@@ -28,9 +34,6 @@ abstract class AnalogRadio {
 
   /// Internal signal strength. From 0 to 1
   double internalSignalStrength = 1;
-
-  /// Applied filter
-  FilterType filter = FilterType.allpass;
 
   /// The radio is playing
   bool get playing;