+++ lib/src/filter_type.dart
@@ -1,2 +1,0 @@
-/// Sound filter
-enum FilterType { bandpass, allpass }