+++ web/main.dart
@@ -1,4 +1,7 @@
 import 'dart:html';
+import 'dart:math';
+import 'dart:typed_data';
+import 'dart:web_audio';
 
 import 'package:analog_radio/analog_radio.dart';
 
@@ -10,8 +13,13 @@ void main() async {
   var $freq = document.querySelector('.freq') as RangeInputElement;
   var $freqValue = document.querySelector('.frequency-value') as Element;
   var $stations = document.querySelector('.stations') as TextAreaElement;
+  var $visual = document.getElementById('visual') as CanvasElement;
+  late AnalyserNode analyser;
 
-  var radio = AnalogRadio();
+  var radio = AnalogRadio(setCustomNode: (context) {
+    analyser = context.createAnalyser();
+    return analyser;
+  });
   var stations = Stations();
 
   $freq.setAttribute('min', stations.frequencyMin);
@@ -21,6 +29,8 @@ void main() async {
     $freqValue.innerText = $freq.value!;
     var freq = int.tryParse($freq.value!) ?? 0;
     var station = stations.getStationByFreq(freq);
+
+    radio.tuning();
 
     if (station == null) {
       radio.signalStrength = 0;
@@ -37,12 +47,12 @@ void main() async {
 
   var volumeHandler = (Event event) {
     var val = double.parse((event.target as InputElement).value!);
-    if (val == 0) {
+    if (val == -30) {
       radio.turnOff();
     } else {
       radio.turnOn();
     }
-    radio.volume = val;
+    radio.volume = pow(10, (val / 20)).toDouble();
   };
 
   void drawStationsFrequency(List<Station> stations) {
@@ -70,7 +80,7 @@ void main() async {
     drawStationsFrequency(stations.stations);
   };
 
-  $volume.value = '0';
+  $volume.value = '-30';
 
   var initialFreq = window.location.hash.replaceAll('#', '');
 
@@ -86,4 +96,44 @@ void main() async {
   $volume.onInput.listen(volumeHandler);
   $freq.onInput.listen(freqHandler);
   $stations.onChange.listen(stationsChangeHandler);
+
+  // Draw visualization
+  var WIDTH = 600;
+  var HEIGHT = 200;
+  analyser.fftSize = 2048;
+  var ctx = $visual.getContext('2d') as CanvasRenderingContext2D;
+  ctx.clearRect(0, 0, WIDTH, HEIGHT);
+  var bufferLength = analyser.frequencyBinCount;
+  var dataArray = Uint8List(bufferLength!);
+
+  void drawVisual(_) {
+    analyser.getByteTimeDomainData(dataArray);
+    ctx.fillStyle = "rgb(33, 33, 33)";
+    ctx.fillRect(0, 0, WIDTH, HEIGHT);
+    ctx.lineWidth = 2;
+    ctx.strokeStyle = "rgb(220, 220, 220)";
+    ctx.beginPath();
+    var sliceWidth = WIDTH / bufferLength;
+    double x = 0;
+
+    for (var i = 0; i < bufferLength; i++) {
+      var v = dataArray[i] / 128.0;
+      var y = HEIGHT - v * (HEIGHT / 2);
+
+      if (i == 0) {
+        ctx.moveTo(x, y);
+      } else {
+        ctx.lineTo(x, y);
+      }
+
+      x += sliceWidth;
+    }
+
+    ctx.lineTo(WIDTH, HEIGHT / 2);
+    ctx.stroke();
+
+    window.requestAnimationFrame(drawVisual);
+  }
+
+  window.requestAnimationFrame(drawVisual);
 }